Implementação Refresh Token

+++ cryptoexchange/src/main/java/com/cryptoexchange/cryptoexchange/models/Role.java
@@ -1,7 +1,13 @@
 package com.cryptoexchange.cryptoexchange.models;
 
-import javax.persistence.*;
-
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
 
 // Cria a tabela roles no BD
 @Entity
@@ -9,7 +15,7 @@ import javax.persistence.*;
 public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Enumerated(EnumType.STRING)
     @Column(length = 25)
@@ -25,11 +31,11 @@ public class Role {
 
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
